requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
djangorestframework==3.7.3
idna==2.6
lxml==4.1.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
